WIP elastic 6.x
Only one type per index is permitted.
Opted to fake multiple types by adding it as a field in the mapping.

Id's are no longer unique per type, but per index. So we prepend the
type ( $type-$id ). It's the user's responsibility to make sure the
id he returns in getElasticId($type) is unique in the entire index.