Refactor: Allow multiple documents per entity
We required a single entity to have multiple documents for a doctype.
Since we had to change the design we took the liberty to also require
the use of a service to generate the elastic array.

We really try to avoid \Drupal:: calls in our models ( wieni/wmmodel )
but the previous design didn't allow for this.

----

Might be a bit harder to wrap your head around now 🥺

- DocumentInterface renamed to ElasticEntityInterface and simplified
  - Contains only getElasticTypes() and getElasticDocumentCollection($type)
- DocumentCollectionInterface is new and responsible for creating the
  elastic array and for providing the elastic ids ( yeah, multiple ).
  A DocumentCollection should be aware of the doctype it's managing (
  so don't use one DocumentCollection for 'page' and 'related' doctypes ).
- DocumentCollectionInterface::getElasticIds() returns an ARRAY of ids.
  The module will keep track of new/removed ids and create/remove when
  needed.
- Previously this module only deleted documents in the 'page' when an
  entity is removed. In this version it also deletes for every doctype
  returned by ElasticEntityInterface::getElasticTypes()
 - Updated README